Enforce NOMINMAX macro everywhere in Mozc.
With recent toolchains such as Visual C++ 2013 and Windows 8.1 SDK, we
are finally able to define NOMINMAX globally and unexceptionally without
breaking dependent libraries.  Now that including Windows.h no longer
defines min/max as a macro, we can remove per-file "#undef min" and
"#undef max" safely.

BUG=
TEST=unittest
REF_BUG=19010851
REF_CL=86940185

+++ src/win32/ime/ime_ui_window.cc
@@ -39,6 +39,7 @@
 #include <atlmisc.h>
 #include <strsafe.h>
 
+#include <algorithm>
 #include <memory>
 
 #include "base/const.h"
@@ -349,7 +350,7 @@ class LangBarCallbackImpl : public LangBarCallback {
 
   virtual ULONG AddRef() {
     const LONG count = ::InterlockedIncrement(&reference_count_);
-    return max(count, 0);
+    return static_cast<ULONG>(max(count, static_cast<LONG>(0)));
   }
 
   virtual ULONG Release() {
